Restoring some structure
