+++ data_as_code/__init__.py
@@ -1,4 +1,6 @@
+from data_as_code._misc import ingredient
 from data_as_code._recipe import Recipe, Keep
-from data_as_code.step import Step, source_local, source_http, ingredient
+from data_as_code._step import Step
+from data_as_code.premade import source_local, source_http
 
 __version__ = '0.0.0'
