+++ data_as_code/_misc.py
@@ -1,3 +1,31 @@
+from pathlib import Path
+
 source = 'source'
 intermediary = 'intermediary'
 product = 'product'
+
+
+class _Ingredient:
+    def __init__(self, step_name: str):
+        self.step_name = step_name
+
+
+def ingredient(step: str) -> Path:
+    """
+    Prepare step ingredient
+
+    Use the metadata from a previously executed step as an ingredient for
+    another step. This function is a wrapper to allowing passing the results of
+    a previous step directly to the next, while still allowing context hints to
+    function appropriately.
+
+    The typehint says that the return is a :class:`pathlib.Path` object, but
+    that is a lie; the return is actually a semi-private Ingredient class. This
+    mismatch is done intentionally to allow all ingredients in a step to be
+    identified without first knowing the names of the attributes that they will
+    be assigned to. Once the ingredients are captured, the attribute is
+    reassigned to the path attribute for the ingredient, allowing the path to
+    be called directly from inside the :class:`Step.instructions`
+    """
+    # noinspection PyTypeChecker
+    return _Ingredient(step)
